extract adjustmentRanges and defaultRateProfileIndex from profile_t. adjustmentRanges is wrapped in adjustmentProfile_t.  extracted rate profile code from config.c into rate_profile.  extracted adjustments code into rc_adjustments.c.
The latter two extractions of code will help with better unit test
isolation and code-hunting too.  rc_controls.c was getting a bit busy.
