+++ src/main/config/config_profile.h
@@ -18,7 +18,6 @@
 #pragma once
 
 typedef struct profile_s {
-    uint8_t defaultRateProfileIndex;
 
     int16_t mag_declination;                // Get your magnetic decliniation from here : http://magnetic-declination.com/
                                             // For example, -6deg 37min, = -637 Japan, format is [sign]dddmm (degreesminutes) default is zero.
@@ -27,10 +26,7 @@ typedef struct profile_s {
     barometerConfig_t barometerConfig;
 #endif
 
-
     modeActivationCondition_t modeActivationConditions[MAX_MODE_ACTIVATION_CONDITION_COUNT];
-
-    adjustmentRange_t adjustmentRanges[MAX_ADJUSTMENT_RANGE_COUNT];
 
     // Radio/ESC-related configuration
