+++ src/main/config/parameter_group_ids.h
@@ -51,6 +51,8 @@
 #define PG_HOTT_TELEMETRY_CONFIG 33
 #define PG_NAVIGATION_CONFIG 34
 #define PG_ACCELEROMETER_CONFIG 35
+#define PG_RATE_PROFILE_SELECTION 36
+#define PG_ADJUSTMENT_PROFILE 37
 
 // Driver configuration
 #define PG_DRIVER_PWM_RX_CONFIG 100
