+++ src/main/io/display.c
@@ -41,6 +41,7 @@
 
 #ifdef DISPLAY
 
+#include "io/rate_profile.h"
 #include "io/rc_controls.h"
 
 #include "sensors/battery.h"
