+++ src/main/io/rate_profile.h
@@ -1,0 +1,48 @@
+/*
+ * This file is part of Cleanflight.
+ *
+ * Cleanflight is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * Cleanflight is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with Cleanflight.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#pragma once
+
+typedef struct controlRateConfig_s {
+    uint8_t rcRate8;
+    uint8_t rcExpo8;
+    uint8_t thrMid8;
+    uint8_t thrExpo8;
+    uint8_t rates[3];
+    uint8_t dynThrPID;
+    uint8_t rcYawExpo8;
+    uint16_t tpa_breakpoint;                // Breakpoint at which TPA is activated
+} controlRateConfig_t;
+
+#define MAX_CONTROL_RATE_PROFILE_COUNT 3
+PG_DECLARE_ARR(controlRateConfig_t, MAX_CONTROL_RATE_PROFILE_COUNT, controlRateProfiles);
+
+extern controlRateConfig_t *currentControlRateProfile;
+
+typedef struct rateProfileSelection_s {
+    uint8_t defaultRateProfileIndex;
+} rateProfileSelection_t;
+
+PG_DECLARE_PROFILE(rateProfileSelection_t, rateProfileSelection);
+
+void setControlRateProfile(uint8_t profileIndex);
+uint8_t getCurrentControlRateProfile(void);
+controlRateConfig_t *getControlRateConfig(uint8_t profileIndex);
+void configureRateProfileSelection(uint8_t profileIndex, uint8_t rateProfileIndex);
+void resetControlRateConfig(controlRateConfig_t *controlRateConfig);
+
+void activateControlRateConfig(void);
