+++ src/main/io/rc_controls.h
@@ -131,22 +131,6 @@ typedef struct modeActivationCondition_s {
 
 #define IS_RANGE_USABLE(range) ((range)->startStep < (range)->endStep)
 
-typedef struct controlRateConfig_s {
-    uint8_t rcRate8;
-    uint8_t rcExpo8;
-    uint8_t thrMid8;
-    uint8_t thrExpo8;
-    uint8_t rates[3];
-    uint8_t dynThrPID;
-    uint8_t rcYawExpo8;
-    uint16_t tpa_breakpoint;                // Breakpoint at which TPA is activated
-} controlRateConfig_t;
-
-#define MAX_CONTROL_RATE_PROFILE_COUNT 3
-PG_DECLARE_ARR(controlRateConfig_t, MAX_CONTROL_RATE_PROFILE_COUNT, controlRateProfiles);
-
-extern controlRateConfig_t *currentControlRateProfile;
-
 extern int16_t rcCommand[4];
 
 typedef struct rcControlsConfig_s {
@@ -179,117 +163,6 @@ void processRcStickPositions(rxConfig_t *rxConfig, throttleStatus_e throttleStat
 
 void updateActivatedModes(modeActivationCondition_t *modeActivationConditions);
 
-#define PID_MIN      0
-#define PID_MAX      200
-#define PID_F_MIN    0
-#define PID_F_MAX    100
-#define RC_RATE_MIN  0
-#define RC_RATE_MAX  250
-#define EXPO_MIN     0
-#define EXPO_MAX     100
-
-typedef enum {
-    ADJUSTMENT_NONE = 0,
-    ADJUSTMENT_RC_RATE,
-    ADJUSTMENT_RC_EXPO,
-    ADJUSTMENT_THROTTLE_EXPO,
-    ADJUSTMENT_PITCH_ROLL_RATE,
-    ADJUSTMENT_YAW_RATE,
-    ADJUSTMENT_PITCH_ROLL_P,
-    ADJUSTMENT_PITCH_ROLL_I,
-    ADJUSTMENT_PITCH_ROLL_D,
-    ADJUSTMENT_YAW_P,
-    ADJUSTMENT_YAW_I,
-    ADJUSTMENT_YAW_D,
-    ADJUSTMENT_RATE_PROFILE,
-    ADJUSTMENT_PITCH_RATE,
-    ADJUSTMENT_ROLL_RATE,
-    ADJUSTMENT_PITCH_P,
-    ADJUSTMENT_PITCH_I,
-    ADJUSTMENT_PITCH_D,
-    ADJUSTMENT_ROLL_P,
-    ADJUSTMENT_ROLL_I,
-    ADJUSTMENT_ROLL_D,
-    ADJUSTMENT_ALT_P,
-    ADJUSTMENT_ALT_I,
-    ADJUSTMENT_ALT_D,
-    ADJUSTMENT_VEL_P,
-    ADJUSTMENT_VEL_I,
-    ADJUSTMENT_VEL_D,
-    ADJUSTMENT_MAG_P,
-    ADJUSTMENT_POS_P,
-    ADJUSTMENT_POS_I,
-    ADJUSTMENT_POSR_P,
-    ADJUSTMENT_POSR_I,
-    ADJUSTMENT_POSR_D,
-    ADJUSTMENT_NAVR_P,
-    ADJUSTMENT_NAVR_I,
-    ADJUSTMENT_NAVR_D,
-    ADJUSTMENT_LEVEL_P,
-    ADJUSTMENT_LEVEL_I,
-    ADJUSTMENT_LEVEL_D,
-
-} adjustmentFunction_e;
-
-#define ADJUSTMENT_FUNCTION_COUNT 39
-
-typedef enum {
-    ADJUSTMENT_MODE_STEP,
-    ADJUSTMENT_MODE_SELECT
-} adjustmentMode_e;
-
-typedef struct adjustmentStepConfig_s {
-    uint8_t step;
-} adjustmentStepConfig_t;
-
-typedef struct adjustmentSelectConfig_s {
-    uint8_t switchPositions;
-} adjustmentSelectConfig_t;
-
-typedef union adjustmentConfig_u {
-    adjustmentStepConfig_t stepConfig;
-    adjustmentSelectConfig_t selectConfig;
-} adjustmentData_t;
-
-typedef struct adjustmentConfig_s {
-    uint8_t adjustmentFunction;
-    uint8_t mode;
-    adjustmentData_t data;
-} adjustmentConfig_t;
-
-typedef struct adjustmentRange_s {
-    // when aux channel is in range...
-    uint8_t auxChannelIndex;
-    channelRange_t range;
-
-    // ..then apply the adjustment function to the auxSwitchChannel ...
-    uint8_t adjustmentFunction;
-    uint8_t auxSwitchChannelIndex;
-
-    // ... via slot
-    uint8_t adjustmentIndex;
-} adjustmentRange_t;
-
-#define ADJUSTMENT_INDEX_OFFSET 1
-
-typedef struct adjustmentState_s {
-    uint8_t auxChannelIndex;
-    adjustmentConfig_t config;
-    uint32_t timeoutAt;
-    adjustmentRange_t *range;
-} adjustmentState_t;
-
-
-#ifndef MAX_SIMULTANEOUS_ADJUSTMENT_COUNT
-#define MAX_SIMULTANEOUS_ADJUSTMENT_COUNT 4 // enough for 4 x 3position switches / 4 aux channel
-#endif
-
-#define MAX_ADJUSTMENT_RANGE_COUNT 12 // enough for 2 * 6pos switches.
-
-void resetAdjustmentStates(void);
-void configureAdjustmentState(adjustmentRange_t *adjustmentRange);
-void updateAdjustmentStates(adjustmentRange_t *adjustmentRanges);
-void processRcAdjustments(controlRateConfig_t *controlRateConfig, rxConfig_t *rxConfig);
 
 bool isUsingSticksForArming(void);
 
@@ -297,3 +170,5 @@ int32_t getRcStickDeflection(int32_t axis, uint16_t midrc);
 bool isModeActivationConditionPresent(modeActivationCondition_t *modeActivationConditions, boxId_e modeId);
 
 void useRcControlsConfig(modeActivationCondition_t *modeActivationConditions);
+
+bool isRangeActive(uint8_t auxChannelIndex, channelRange_t *range);
