+++ src/main/io/rc_curves.c
@@ -21,6 +21,7 @@
 #include "config/parameter_group.h"
 
 #include "rx/rx.h"
+#include "io/rate_profile.h"
 #include "io/rc_controls.h"
 #include "io/escservo.h"
 
@@ -31,34 +32,34 @@ int16_t lookupYawRC[YAW_LOOKUP_LENGTH];     // lookup table for expo & RC rate Y
 int16_t lookupThrottleRC[THROTTLE_LOOKUP_LENGTH];   // lookup table for expo & mid THROTTLE
 
 
-void generatePitchRollCurve(controlRateConfig_t *controlRateConfig)
+void generatePitchRollCurve()
 {
     uint8_t i;
 
     for (i = 0; i < PITCH_LOOKUP_LENGTH; i++)
-        lookupPitchRollRC[i] = (2500 + controlRateConfig->rcExpo8 * (i * i - 25)) * i * (int32_t) controlRateConfig->rcRate8 / 2500;
+        lookupPitchRollRC[i] = (2500 + currentControlRateProfile->rcExpo8 * (i * i - 25)) * i * (int32_t) currentControlRateProfile->rcRate8 / 2500;
 }
 
-void generateYawCurve(controlRateConfig_t *controlRateConfig)
+void generateYawCurve()
 {
     uint8_t i;
 
     for (i = 0; i < YAW_LOOKUP_LENGTH; i++)
-        lookupYawRC[i] = (2500 + controlRateConfig->rcYawExpo8 * (i * i - 25)) * i / 25;
+        lookupYawRC[i] = (2500 + currentControlRateProfile->rcYawExpo8 * (i * i - 25)) * i / 25;
 }
 
-void generateThrottleCurve(controlRateConfig_t *controlRateConfig, escAndServoConfig_t *escAndServoConfig)
+void generateThrottleCurve()
 {
     uint8_t i;
 
     for (i = 0; i < THROTTLE_LOOKUP_LENGTH; i++) {
-        int16_t tmp = 10 * i - controlRateConfig->thrMid8;
+        int16_t tmp = 10 * i - currentControlRateProfile->thrMid8;
         uint8_t y = 1;
         if (tmp > 0)
-            y = 100 - controlRateConfig->thrMid8;
+            y = 100 - currentControlRateProfile->thrMid8;
         if (tmp < 0)
-            y = controlRateConfig->thrMid8;
-        lookupThrottleRC[i] = 10 * controlRateConfig->thrMid8 + tmp * (100 - controlRateConfig->thrExpo8 + (int32_t) controlRateConfig->thrExpo8 * (tmp * tmp) / (y * y)) / 10;
-        lookupThrottleRC[i] = escAndServoConfig->minthrottle + (int32_t) (escAndServoConfig->maxthrottle - escAndServoConfig->minthrottle) * lookupThrottleRC[i] / 1000; // [MINTHROTTLE;MAXTHROTTLE]
+            y = currentControlRateProfile->thrMid8;
+        lookupThrottleRC[i] = 10 * currentControlRateProfile->thrMid8 + tmp * (100 - currentControlRateProfile->thrExpo8 + (int32_t) currentControlRateProfile->thrExpo8 * (tmp * tmp) / (y * y)) / 10;
+        lookupThrottleRC[i] = escAndServoConfig.minthrottle + (int32_t) (escAndServoConfig.maxthrottle - escAndServoConfig.minthrottle) * lookupThrottleRC[i] / 1000; // [MINTHROTTLE;MAXTHROTTLE]
     }
 }
