+++ src/test/unit/pid_unittest.cc
@@ -36,6 +36,7 @@ extern "C" {
 
     #include "rx/rx.h"
     #include "io/rc_controls.h"
+    #include "io/rate_profile.h"
 
     #include "flight/pid.h"
     #include "flight/imu.h"
